# atodo-api-spec.yaml
...
openapi: 3.0.3
info:
  title: A-To-Do API
  description: >-
    Backend API for the A-To-Do to-do list application. Implementing this
    spec turns the app from a single-browser, localStorage-only prototype
    into a real multi-tenant service: every endpoint below is scoped to the
    authenticated account (via its bearer token), so two people logging into
    two different accounts genuinely see two different task lists, rather
    than the one shared list the pre-backend prototype used.

    All endpoints are prefixed with `/atodo/v1`. The bearer token is an
    opaque, server-issued credential (a real signed JWT is the intended
    implementation) that embeds at least `sub` (the account id) and the
    account's current `subscription` claim, so the client can read
    subscription status without an extra round trip -- see the `Token`
    schema. Whenever an endpoint changes something the token's claims
    reflect (starting/cancelling/scheduling deletion of a subscription), it
    returns a freshly issued token in place of the old one; the client is
    expected to discard the old one immediately.

    Every error response uses the same `Error` schema (`{ code, message }`).
    `code` is a stable, machine-readable identifier the client branches on
    (e.g. to show a specific translated message); `message` is a
    human-readable fallback, not meant to be shown as-is in a finished UI.
  version: "1.0.0"
...
servers:
  - url: "{baseUrl}/atodo/v1"
    variables:
...
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new (unverified) account
      description: >-
        Creates a pending registration and emails a verification link to
        `email` (server-side email delivery; the client never sees the
        token). Registering again with an email that already has a pending
        (not yet verified) registration re-sends a fresh link rather than
        erroring, since the first one may have gone missing or expired.
        Never reveals whether an email is already a *verified* account
        (that's `EMAIL_TAKEN`) vs. never registered at all -- see `login`'s
        own description for the same principle.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        "202":
          description: Verification email sent (or re-sent).
        "400":
          description: Malformed email or a password under 8 characters.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: INVALID_EMAIL, message: "Enter a valid email address." }
        "409":
          description: A verified account already exists with this email.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: EMAIL_TAKEN, message: "An account with that email already exists." }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Consume an email verification link's token
      description: >-
        Promotes a pending registration to a real, loggable-into account.
        Single-use: the token is consumed as soon as it's looked up here,
        whether or not it was still within its validity window, so
        revisiting the same link a second time always reports
        `INVALID_TOKEN` rather than re-verifying or re-expiring.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Verified.
          content:
            application/json:
              schema:
                type: object
                properties:
                  email: { type: string, format: email }
        "400":
          description: The token doesn't match any pending registration (already used, or never existed).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: INVALID_TOKEN, message: "That verification link is invalid or has already been used." }
        "410":
          description: The token matched a pending registration, but its validity window (6 hours) has passed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: EXPIRED, message: "That verification link has expired. Please register again." }
  /auth/login:
    post:
      tags: [Auth]
      summary: Exchange email + password for a bearer token
      description: >-
        Never distinguishes "no such account" from "wrong password" in the
        response (both are `INVALID_CREDENTIALS`) -- only an account that
        genuinely exists but hasn't verified its email yet gets its own
        `EMAIL_NOT_VERIFIED` code, so a mistyped email a visitor never
        actually registered doesn't leak which emails are/aren't
        registered. Also where the 12-month-inactivity and
        scheduled-account-deletion policies (see the Privacy Policy and
        `DELETE /users/me/schedule-deletion`) actually take effect: logging
        into an account that's now due for either kind of deletion deletes
        it right here, before a token is ever issued for it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { $ref: "#/components/schemas/Token" }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Wrong email or password.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: INVALID_CREDENTIALS, message: "Incorrect email or password." }
        "403":
          description: The account exists but its email hasn't been verified yet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: EMAIL_NOT_VERIFIED, message: "That email hasn't been verified yet." }
        "410":
          description: The account existed but has just been deleted (inactivity or scheduled deletion) as part of processing this request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                inactivity:
                  value: { code: ACCOUNT_EXPIRED_INACTIVITY, message: "This account was automatically deleted after 12 months of inactivity." }
                scheduled:
                  value: { code: ACCOUNT_DELETED_SCHEDULED, message: "This account was deleted, as scheduled, once its subscription ended." }
  /auth/me:
    get:
      tags: [Auth]
      summary: Resolve the current bearer token into the live account it belongs to
      description: >-
        The one endpoint that can tell a still-valid session apart from one
        that isn't anymore -- a token's own claims are a snapshot from
        whenever it was issued, not live truth (see the `Token` schema), so
        the client is expected to call this (not just trust a locally
        decoded token) whenever it needs to know the account is still
        genuinely there, e.g. once per app load. Applies the same
        12-month-inactivity/scheduled-deletion checks `POST /auth/login`
        does, since resuming a stored token is just as much "using the
        account" as a fresh login -- touches the account's `lastActiveAt`
        (server-internal bookkeeping, not part of the `User` schema below,
        since the client never needs to read it) the same way a fresh
        `POST /auth/login` touches `lastLoginAt`, so a long-lived token
        that keeps getting resumed here isn't wrongly treated as abandoned.
      responses:
        "200":
          description: Still a live account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/malformed/expired bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "410":
          description: The account behind this token is gone -- deleted outright (`ACCOUNT_NOT_FOUND`, e.g. from Settings or another session), for 12-month inactivity, or per a completed scheduled deletion. The client must treat any of these as "not logged in" and discard the token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                notFound:
                  value: { code: ACCOUNT_NOT_FOUND, message: "This account no longer exists." }
                inactivity:
                  value: { code: ACCOUNT_EXPIRED_INACTIVITY, message: "This account was automatically deleted after 12 months of inactivity." }
                scheduled:
                  value: { code: ACCOUNT_DELETED_SCHEDULED, message: "This account was deleted, as scheduled, once its subscription ended." }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <token>`, as returned by `POST /auth/login`
        or any of the subscription-mutating endpoints. See the `Token`
        schema for what it's expected to carry.
  schemas:
    EpochMillis:
      type: integer
      format: int64
      description: >-
        Milliseconds since the Unix epoch (`Date.now()`/`new Date(ms)` on the
        client), not an RFC 3339 string -- deliberately, since every one of
        these fields is used in client-side arithmetic/sorting (elapsed
        time, "is this still active", "which of these was created first",
        ...) and epoch millis need no parsing to do that. `dueDate`/
        `endDate`/occurrence-date fields elsewhere are a different thing --
        plain `YYYY-MM-DD` strings with no time component, compared as
        strings, so those stay `format: date`.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable -- the client branches on this, not `message`.
        message:
          type: string
          description: Human-readable fallback; a finished client should show its own translated copy per `code` instead.
    Token:
      type: string
      description: >-
        Opaque to the client beyond reading its claims (a real
        implementation: a signed JWT). Expected claims: `sub` (the account
        id) and `subscription` (the same shape as `User.subscription`, a
        snapshot as of when the token was issued -- re-issued by every
        endpoint that changes it, see each one's own description). The
        client never trusts a locally decoded token's claims for anything
        but quick, non-authoritative UI reads (e.g. "should the Subscribe
        button show right now") -- `GET /auth/me` is the actual source of
        truth for whether the account/subscription are still genuinely in
        that state.
    Background:
      type: object
      nullable: true
      description: The Settings background picker's currently-applied Unsplash photo, or null for none.
      properties:
        regularUrl: { type: string, format: uri }
        thumbUrl: { type: string, format: uri }
        photographerName: { type: string }
        photographerUrl: { type: string, format: uri }
        photoLink: { type: string, format: uri }
    Subscription:
      type: object
      nullable: true
      description: null means the account has never subscribed (the "Free" plan).
      properties:
        id: { type: string }
        plan:
          type: string
          enum: [trial, pro]
        billingInterval:
          type: string
          enum: [monthly, annual]
          nullable: true
          description: null for a trial.
        startedAt: { $ref: "#/components/schemas/EpochMillis" }
        expiresAt:
          allOf: [{ $ref: "#/components/schemas/EpochMillis" }]
          description: >-
            The single source of truth for whether the subscription is
            currently active -- always `now < expiresAt`, never a separately
            cached boolean, so an expiry is correctly detected the instant
            it passes rather than only once something happens to re-check it.
        cancelAtPeriodEnd:
          type: boolean
          description: True once the subscription won't renew (see POST /subscriptions/cancel and POST /users/me/schedule-deletion) -- access is untouched until expiresAt regardless.
        scheduledDeletion:
          type: boolean
          description: True if the account itself, not just the subscription, is set to be deleted once expiresAt passes -- see POST /users/me/schedule-deletion.
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        nickname: { type: string }
        avatar:
          type: string
          nullable: true
          description: A data URL, or null for the initials-only fallback avatar.
        timeFormat: { type: string, enum: ["12", "24"] }
        background: { $ref: "#/components/schemas/Background" }
        language:
          type: string
          enum: [en, hr]
          nullable: true
          description: null means never chosen or IP-detected yet -- the client detects and PATCHes a real value the first time this is null right after login.
        activeTaskId:
          type: string
          nullable: true
          description: Which task (by Task.id) is currently focused/being timed, if any.
        activeOccurrenceDate:
          type: string
          format: date
          nullable: true
          description: Which occurrence of activeTaskId, for a recurring task that can show more than one row at once. Always null when activeTaskId is null.
        todoViewMode:
          type: string
          enum: [pending, next-recurrence, all]
        subscription: { $ref: "#/components/schemas/Subscription" }
...